rest/requirements.txt
django==4.2.2
gunicorn==20.1.0
psycopg2==2.9.6
coverage==7.2.7
django-extensions==3.2.3
pydotplus==2.0.2
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-ninja-jwt==5.2.5
orjson==3.9.1
six==1.16.0
